pyproject.toml
[project]
name = "arsip-baranahan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.0",
    "fastapi[standard]>=0.115.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.10.1",
    "python-jose[cryptography]>=3.3.0",
    "ruff>=0.8.1",
]
